Removed uneeded folders/files
